Changes to Xavier Initialization

+++ assignment2_old/q2_initialization.py
@@ -25,8 +25,8 @@ def xavier_weight_init():
             out: tf.Tensor of specified shape sampled from the Xavier distribution.
         """
         ### YOUR CODE HERE
-        epsilon = tf.sqrt(tf.divide(6, tf.reduce_sum(shape)))
-        out = tf.random_uniform(shape, -epsilon, epsilon, dtype=tf.float64)
+        epsilon = tf.sqrt(tf.cast(tf.divide(6, tf.reduce_sum(shape)), dtype=tf.float32))
+        out = tf.random_uniform(shape, -epsilon, epsilon)
         ### END YOUR CODE
         return out
     # Returns defined initializer function.
